pyproject.toml
[project]
name = "rosetta-sdl"
version = "0.1.0"
description = "Rosetta SDL — Translate business language into data insights. A domain-agnostic semantic data layer with Neo4j ontology for AWS data lakes."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "uvicorn>=0.34.0",
    "neo4j>=5.25.0",
    "boto3>=1.35.0",
    "sqlglot>=26.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "pydantic>=2.0",
    "pyjwt[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "testcontainers[neo4j]>=4.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
